Trying to fix: Logos are largely (or for kandji totally) invisible in dark mode in README.md
An edit that should work, per https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax#specifying-the-theme-an-image-is-shown-to and I think it's an improvement but doesn't actually seem to be working right.  I don't get why there are 3 image URLs.  At least in preview, it seems to be using the 3rd URL.  I expected it to be using the first.